FROMGIT: thunderbolt: Enable/disable sideband depending on USB4 port offline mode
When USB4 port is in offline mode (this mean there is no device
attached) we want to keep the sideband up to make it possible to
communicate with the retimers. In the same way there is no need to
enable sideband transactions when the USB4 port is not offline as they
are already up.

For this reason make the enabling/disabling depend on the USB4 port
offline status.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 87200371817ec6e48e42ef2f03756268661a8815
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git/
next)

BUG=b:279527866
TEST=Verify Retimer firmware update in NDA.

Change-Id: I7b6272afdd51b36d09ff77e26507c8028458e15c
Signed-off-by: Utkarsh Patel <utkarsh.h.patel@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4610850
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Prashant Malani <pmalani@chromium.org>